backend/pyproject.toml
[project]
name = "todo-backend"
version = "1.0.0"
description = "Todo Web Application Backend - FastAPI with SQLModel"
requires-python = ">=3.13"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.115.0",
    "sqlmodel>=0.0.24",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.28.0",
    "uvicorn[standard]>=0.32.0",
    "asyncpg>=0.30.0",
    "python-multipart>=0.0.20",
    "structlog>=24.1.0",
    "python-json-logger>=2.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "black>=24.0.0",
    "ruff>=0.8.0",
    "mypy>=1.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py313"
line-length = 100
select = ["E", "F", "I", "UP", "W"]
ignore = ["E501"]

[tool.black]
target-version = ["py313"]
line-length = 100

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
